Added integration test for TPM.
