+++ ee/orbit/pkg/securehw/securehw_stub.go
@@ -1,5 +1,4 @@
 //go:build !linux
-// +build !linux
 
 package securehw
